Split the dependencies
