+++ src/rt/mod.rs
@@ -5,7 +5,11 @@
 mod futures;
 pub use futures::FuturesIo;
 
-#[cfg(feature = "smol")]
-mod smol;
-#[cfg(feature = "smol")]
-pub use smol::{SmolExecutor, SmolTimer};
+#[cfg(feature = "async-executor")]
+mod async_executor;
+#[cfg(feature = "async-io")]
+mod async_io;
+#[cfg(feature = "async-executor")]
+pub use async_executor::SmolExecutor;
+#[cfg(feature = "async-io")]
+pub use async_io::SmolTimer;
